Moved merge-sort to Typescript
